+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - isArray validator can now validate items with validators and multiple types.
-- `@Controller` and all route decorators do accept `RequestHandler` as middlewares.
+- `@Controller` and all route decorators do optionally accept `RequestHandler` as middlewares.
 - Example code for simple middleware.
 - Example code for passportJS middleware.